Add fastrand and use chrono for uptime
Compute uptime using chrono::Local DateTime and return hours as f64.
Initialize START_TIME with a DateTime instead of Instant. Update
HealthCheckResponse.uptime_hrs to f64. Also add fastrand dependency and
mark generate_random_vectors in benches as #[inline].

+++ src/server/dto.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 pub struct HealthCheckResponse {
     pub status: String,
     pub service: String,
-    pub uptime_hrs: f32,
+    pub uptime_hrs: f64,
 }
 
 /// Request DTO for database creation
